+++ HISTORY.md
@@ -1,5 +1,10 @@
 # History
 
+## 0.1.1 (2021-11-10')
+
+* Serialize JSON as compact by default
+* add `.to_int()` to convert a BinaPy to an int.
+
 ## 0.1.0 (2021-07-08)
 
 * First release on PyPI.